Introduce `geometry.svg` package
