+++ app/src/jvmMain/kotlin/diy/lingerie/geometry/svg/PrimitiveCurve.kt
@@ -1,0 +1,56 @@
+package diy.lingerie.geometry.svg
+
+import diy.lingerie.geometry.LineSegment
+import diy.lingerie.geometry.curves.PrimitiveCurve
+import diy.lingerie.geometry.curves.bezier.BezierCurve
+import diy.lingerie.geometry.curves.bezier.BezierCurve.Edge
+import diy.lingerie.geometry.splines.Spline
+import diy.lingerie.simple_dom.svg.SvgPath
+import diy.lingerie.simple_dom.svg.SvgPath.Segment
+import diy.lingerie.simple_dom.svg.SvgShape
+
+fun BezierCurve.toSvgPath(
+    stroke: SvgShape.Stroke = SvgShape.Stroke.default,
+): SvgPath = SvgPath(
+    stroke = stroke,
+    segments = listOf(
+        SvgPath.Segment.MoveTo(
+            targetPoint = start,
+        ),
+    ) + toSvgBezierSegment(),
+)
+
+fun PrimitiveCurve.toSvgSegment(): SvgPath.Segment = when (this) {
+    is LineSegment -> SvgPath.Segment.LineTo(
+        finalPoint = end,
+    )
+
+    is BezierCurve -> toSvgBezierSegment()
+
+    else -> throw UnsupportedOperationException("Unsupported segment curve: $this")
+}
+
+fun BezierCurve.toSvgBezierSegment(): SvgPath.Segment.CubicBezierCurveTo = SvgPath.Segment.CubicBezierCurveTo(
+    controlPoint1 = firstControl,
+    controlPoint2 = secondControl,
+    finalPoint = end,
+)
+
+fun Segment.CurveSegment.toEdge(): PrimitiveCurve.Edge = when (this) {
+    is Segment.LineTo -> LineSegment.Edge
+
+    is Segment.CubicBezierCurveTo -> Edge(
+        firstControl = controlPoint1,
+        secondControl = controlPoint2,
+    )
+
+    is Segment.SmoothCubicBezierCurveTo -> TODO()
+}
+
+fun Segment.CurveSegment.toLink(): Spline.Link {
+    val edge = toEdge()
+
+    return edge.semiBind(
+        end = finalPoint,
+    )
+}
